// services/messagebox/src/resources/js/store/messageStore/messageStoreMutation.ts
export enum MessageStoreMutation {
    MARK_MESSAGE_READ = 'MARK_MESSAGE_READ',
    RESET_CURRENT_MESSAGE = 'RESET_CURRENT_MESSAGE',
    SET_CURRENT_MESSAGE = 'SET_CURRENT_MESSAGE',
    SET_LIST = 'SET_LIST',
    SET_ERROR = 'SET_ERROR',
    SET_CURRENT_MESSAGE_LOADED = 'SET_CURRENT_MESSAGE_LOADED',
}